+++ src/main/java/nextstep/courses/domain/course/Course.java
@@ -1,4 +1,7 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.course;
+
+import nextstep.courses.domain.BaseTime;
+import nextstep.courses.domain.session.Session;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
