+++ src/main/java/nextstep/courses/domain/enrollment/Enrollment.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enrollment;
 
 import nextstep.users.domain.NsUser;
