+++ src/main/java/nextstep/courses/domain/enrollment/Students.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.enrollment;
 
 import nextstep.courses.AlreadyEnrolledException;
 import nextstep.courses.ExceedMaxEnrollmentException;
