+++ src/main/java/nextstep/courses/domain/payment/FreePaymentStrategy.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.payment;
 
 public class FreePaymentStrategy implements PaymentStrategy {
     @Override
