+++ src/main/java/nextstep/courses/domain/session/SessionInformation.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public class SessionInformation {
