+++ src/main/java/nextstep/courses/domain/session/SessionPeriod.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import java.time.LocalDate;
 import java.time.Period;
