+++ src/test/java/nextstep/courses/domain/CourseTest.java
@@ -1,5 +1,13 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.course.Course;
+import nextstep.courses.domain.enrollment.Enrollment;
+import nextstep.courses.domain.image.Image;
+import nextstep.courses.domain.payment.FreePaymentStrategy;
+import nextstep.courses.domain.session.Session;
+import nextstep.courses.domain.session.SessionInformation;
+import nextstep.courses.domain.session.SessionPeriod;
+import nextstep.courses.domain.session.SessionStatus;
 import nextstep.users.domain.NsUser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
