+++ src/test/java/nextstep/courses/domain/EnrollmentTest.java
@@ -1,5 +1,6 @@
 package nextstep.courses.domain;
 
+import nextstep.courses.domain.enrollment.Enrollment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
